// Front-End/src/app/models/IkGorusmesi.ts
export class IkGorusmesi {
    stepNote: string
    assignee: string
    claimButton: any ='button class="btn btn-fill btn-info btn-sm" type="button" '
    unClaimButton: any ='button class="btn btn-fill btn-info btn-sm" type="button" '
    universalDiv: any ='<div class=" timeline-badge info"><i class=" tim-icons icon-chat-33"></i></div>'
    universalTag: any ='<span class=" badge badge-pill badge-info"> İK Görüşmesi </span>'
    universalButton: any ='<button class="universalButton btn btn-danger btn-sm" id="prensipte_uygun">Prensipte Uygun <i class="tim-icons icon-double-right mb-1"></i></button>'
    universalButton2: any ='<button class="universalButton btn btn-success btn-sm" id="uygun">Uygun <i class="tim-icons icon-double-right mb-1"></i></button>'
    navButton: any ='<button class="navButton btn btn-info btn-sm" id="gorusme_bekliyor">Görüşme Bekleniyor</button>'
    navButton2: any ='<button class="navButton btn btn-info btn-sm" id="yeniden_gorusulecek">Yeniden Görüşülecek</button>'
    navButton3: any ='<button class="navButton btn btn-info btn-sm" id="tecrube_yetersiz">Tecrübe Yetersiz</button>'
    navButton4: any ='<button class="navButton btn btn-info btn-sm" id="sosyal_beceri_yetersiz">Sosyal Beceri Yetersiz</button>'
    navButton5: any ='<button class="navButton btn btn-info btn-sm" id="kara_liste">Kara Liste</button>'
}