// Front-End/src/app/models/TeklifVerilenler.ts
export class TeklifVerilenler {
    stepNote: string
    assignee: string
    claimButton: any ='button class="btn btn-fill btn-offered btn-sm" type="button" '
    unClaimButton: any ='button class="btn btn-fill btn-offered btn-sm" type="button" '
    universalDiv: any ='<div class=" timeline-badge offered"><i class=" tim-icons icon-money-coins"></i></div>'
    universalTag: any ='<span class=" badge badge-pill badge-offered">Teklif Verilenler</span>'
    universalButton: any ='<button class="universalButton btn btn-success btn-sm" id="teklif_gorusmesi_yeniden_yapilsin">Teklif Görüşmesi Yeniden Yapılsın <i class="tim-icons icon-double-right mb-1"></i></button>'
    universalButton2: any ='<button class="universalButton btn btn-linkedin btn-sm" id="musteri_gorusmesi_icin_uygun">Müşteri Görüşmesi İçin Uygun<i class="tim-icons icon-double-right mb-1"></i></button>'
    universalButton3: any ='<button class="universalButton btn btn-light btn-sm" id="kendi_calisan_icin_uygun">Kendi Çalışan İçin Uygun<i class="tim-icons icon-double-right mb-1"></i></button>'
    navButton: any ='<button class="navButton btn btn-offered btn-sm" id="reddetti">Teklifi Reddetti</button>'
}