// Front-End/src/app/models/TeknikGorusme.ts
export class TeknikGorusme {
    stepNote: string
    assignee: string
    claimButton: any ='button class="btn btn-fill btn-danger btn-sm" type="button" '
    unClaimButton: any ='button class="btn btn-fill btn-danger btn-sm" type="button" '
    universalLi: any ='class="timeline-inverted"'
    universalDiv: any ='<div class="timeline-badge danger"><i class="tim-icons icon-puzzle-10"></i></div>'
    universalTag: any ='<span class=" badge badge-pill badge-danger">Teknik Görüşme</span>'
    universalButton: any ='<button class="universalButton btn btn-success btn-sm" id="teknik_olarak_uygun">Teknik Olarak Uygun <i class="tim-icons icon-double-right mb-1"></i></button>'
    universalButton2: any ='<button class="universalButton btn btn-info btn-sm" id="ik_gorusmesi_bastan_yapilsin">İK Görüşmesi Baştan Yapılsın <i class="tim-icons icon-double-right mb-1"></i></button>'
    navButton: any ='<button class="navButton btn btn-danger btn-sm" id="teknik_yetersiz">Teknik Yetersiz</button>'
    navButton2: any ='<button class="navButton btn btn-danger btn-sm" id="tecrube_yetersiz">Tecrübe Yetersiz</button>'
    navButton3: any ='<button class="navButton btn btn-danger btn-sm" id="sosyal_beceri_yetersiz">Sosyal Beceri Yetersiz</button>'
    navButton4: any ='<button class="navButton btn btn-danger btn-sm" id="kara_liste">Kara Liste</button>'
}