// Front-End/src/app/models/MusteriIKGorusme.ts
export class MusteriIKGorusme {
    stepNote: string
    assignee: string
    claimButton: any ='button class="btn btn-fill btn-primary btn-sm" type="button" '
    unClaimButton: any ='button class="btn btn-fill btn-primary btn-sm" type="button" '
    universalLi: any ='class="timeline-inverted"'
    universalDiv: any ='<div class="timeline-badge primary"><i class="tim-icons icon-chat-33"></i></div>'
    universalTag: any ='<span class=" badge badge-pill badge-primary">Müşteri İK Görüşme</span>'
    universalButton: any ='<button class="universalButton btn btn-light btn-sm" id="aday_musteriye_uygun">Aday Müşteriye Uygun<i class="tim-icons icon-double-right mb-1"></i></button>'
    universalButton2: any ='<button class="universalButton btn btn-warning btn-sm" id="musteri_teknik_gorusme">Müşteri Teknik Görüşme<i class="tim-icons icon-double-right mb-1"></i></button>'
    universalButton3: any ='<button class="universalButton btn btn-offered btn-sm" id="teklif_verilenler_yeniden_yapilsin">Teklif Verilenler Yeniden Yapılsın<i class="tim-icons icon-double-right mb-1"></i></button>'
    navButton: any ='<button class="navButton nav btn btn-primary btn-sm" id="aday_cv_uygun_degil">Aday CV Uygun Değil</button>'
}