// Front-End/src/app/models/MusteriTeknikGorusme.ts
export class MusteriTeknikGorusme {
    stepNote: string
    assignee: string
    claimButton: any ='button class="btn btn-fill btn-warning btn-sm" type="button" '
    unClaimButton: any ='button class="btn btn-fill btn-warning btn-sm" type="button" '
    universalDiv: any ='<div class="timeline-badge warning"><i class="tim-icons icon-puzzle-10"></i></div>'
    universalTag: any ='<span class="badge badge-pill badge-warning">Müşteri Teknik Görüşme</span>'
    universalButton: any ='<button class="universalButton btn btn-linkedin btn-sm" id="musteri_ik_gorusmesi_yeniden_yapilsin">Müşteri İK Görüşmesi Yeniden Yapılsın<i class="tim-icons icon-double-right mb-1"></i></button>'
    navButton: any ='<button class="navButton nav btn btn-warning btn-sm" id="adayin_teknigi_yetersiz">Teknik Yetersiz</button>'
    navButton2: any ='<button class="navButton btn btn-warning btn-sm" id="adayin_tecrube_yetersiz">Tecrübe Yetersiz</button>'
    navButton3: any ='<button class="navButton btn btn-warning btn-sm" id="adayin_sosyal_beceri_yetersiz">Sosyal Beceri Yetersiz</button>'
}