// Front-End/src/app/models/AdayIstifasi.ts
export class AdayIstifasi {
    stepNote: string
    assignee: string
    claimButton: any ='button class="btn btn-fill btn-light btn-sm" type="button" '
    unClaimButton: any ='button class="btn btn-fill btn-light btn-sm" type="button" '
    universalLi: any ='class="timeline-inverted"'
    universalDiv: any ='<div class="timeline-badge light"><i class="tim-icons icon-user-run"></i></div>'
    universalTag: any ='<span class=" badge badge-pill badge-light">Aday İstifası</span>'
    universalButton: any ='<button class="universalButton btn btn-reddit btn-sm" id="adayi_denemeye_al">Adayı Denemeye Al<i class="tim-icons icon-double-right mb-1"></i></button>'
    universalButton2: any ='<button class="universalButton btn btn-linkedin btn-sm" id="musteri_ik_gorusmesi_yeniden_yapilsin">Müşteri İK Görüşmesi Yeniden Yapılsın<i class="tim-icons icon-double-right mb-1"></i></button>'
    universalButton3: any ='<button class="universalButton btn btn-offered btn-sm" id="teklif_verilenler_yeniden_yapilsin">Teklif Verilenler Yeniden Yapılsın<i class="tim-icons icon-double-right mb-1"></i></button>'
    navButton: any ='<button class="navButton btn btn-light btn-sm" id="kara_liste">Kara Liste</button>'
}