// Front-End/src/app/models/DenemeSureci.ts
export class DenemeSureci {
    stepNote: string
    assignee: string
    claimButton: any ='button class="btn btn-fill btn-reddit btn-sm" type="button" '
    unClaimButton: any ='button class="btn btn-fill btn-reddit btn-sm" type="button" '
    universalDiv: any ='<div class="timeline-badge reddit"><i class="tim-icons icon-satisfied"></i></div>'
    universalTag: any ='<span class=" badge badge-pill badge-reddit">Deneme Süreci</span>'
    universalButton: any ='<button class="universalButton btn btn-light btn-sm" id="aday_istifasi_yeniden_yapilsin">Aday İstifası Yeniden Yapılsın<i class="tim-icons icon-double-right mb-1"></i></button>'
    navButton: any ='<button class="navButton btn btn-reddit btn-sm" id="calisanlarimiz">Çalışanlarımız</button>'
}