// Front-End/src/app/models/User.ts
export class User {
    key: string
    name: string
    email: string
    password: string
    website: string
    emailVerified: boolean
    photoUrl: string
    linkedin: string
    company: string
    cretedAt: Date
    updatedAt: string
    postacode: number
    phonenumber: number
    position: string
    city: string
    address: string
    country: string
    about: string
    uid: string
    cName: string
}